+++ src/main/java/com/teleclimb/rest/controllers/CompetitionManagementController.java
@@ -18,20 +18,20 @@ public class CompetitionManagementController {
 
     @ApiOperation(value = "Return CUSTOM OBJECT: competition and participants list")
     @GetMapping("/{id}/participants")
-    public CompetitionWithParticipantsList getAllParticipants(@PathVariable Long id) {
+    public CompetitionWithParticipantsList getAllParticipants(@PathVariable Integer id) {
         return additionalService.getAllParticipants(id);
     }
 
     @ApiOperation(value = "Return CUSTOM OBJECT: competition and rounds list")
     @GetMapping("/{id}/rounds")
-    public CompetitionWithRoundsList getAllRounds(@PathVariable Long id) {
+    public CompetitionWithRoundsList getAllRounds(@PathVariable Integer id) {
         return additionalService.getAllRounds(id);
     }
 
 
     @ApiOperation(value = "Generate rounds for competition based on rounds schema. It generate, save to database and return the same as endpoint /competition/{id}/rounds")
     @PostMapping("/{id}/rounds/generate")
-    public CompetitionWithRoundsList generateRounds(@PathVariable Long id) {
+    public CompetitionWithRoundsList generateRounds(@PathVariable Integer id) {
         return additionalService.generateRounds(id);
     }
 }
