+++ src/main/java/com/teleclimb/rest/repositories/CompetitionRepository.java
@@ -3,5 +3,5 @@ package com.teleclimb.rest.repositories;
 import com.teleclimb.rest.entities.Competition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompetitionRepository extends JpaRepository<Competition, Long> {
+public interface CompetitionRepository extends JpaRepository<Competition, Integer> {
 }
