+++ src/main/java/com/teleclimb/rest/repositories/FormulaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormulaRepository extends JpaRepository<Formula, Long> {
+public interface FormulaRepository extends JpaRepository<Formula, Integer> {
 }
