+++ src/main/java/com/teleclimb/rest/services/CompetitionManagementService.java
@@ -2,10 +2,10 @@ package com.teleclimb.rest.services;
 
 import com.teleclimb.responses.error.exception.InternalServerError;
 import com.teleclimb.rest.dto.CompetitionDto;
+import com.teleclimb.rest.dto.ParticipantRawDto;
+import com.teleclimb.rest.dto.RoundRawDto;
 import com.teleclimb.rest.dto.custom.CompetitionWithParticipantsList;
 import com.teleclimb.rest.dto.custom.CompetitionWithRoundsList;
-import com.teleclimb.rest.dto.raw.ParticipantRawDto;
-import com.teleclimb.rest.dto.raw.RoundRawDto;
 import com.teleclimb.rest.entities.Competition;
 import com.teleclimb.rest.entities.Round;
 import com.teleclimb.rest.repositories.CompetitionRepository;
@@ -20,7 +20,7 @@ import java.util.List;
 @Service
 public record CompetitionManagementService(ModelMapper mapper, ParticipantRepository contestantRepo,
                                            RoundRepository roundRepo, CompetitionRepository competitionRepo) {
-    public CompetitionWithParticipantsList getAllParticipants(Long competitionId) {
+    public CompetitionWithParticipantsList getAllParticipants(Integer competitionId) {
         CompetitionWithParticipantsList dto = new CompetitionWithParticipantsList();
 
         dto.setParticipants(contestantRepo.findByCompetitionId(competitionId)
@@ -33,7 +33,7 @@ public record CompetitionManagementService(ModelMapper mapper, ParticipantReposi
         return dto;
     }
 
-    public CompetitionWithRoundsList getAllRounds(Long competitionId) {
+    public CompetitionWithRoundsList getAllRounds(Integer competitionId) {
         CompetitionWithRoundsList dto = new CompetitionWithRoundsList();
 
         dto.setRounds(roundRepo.findByCompetitionId(competitionId)
@@ -47,7 +47,7 @@ public record CompetitionManagementService(ModelMapper mapper, ParticipantReposi
     }
 
 
-    public CompetitionWithRoundsList generateRounds(Long competitionId) {
+    public CompetitionWithRoundsList generateRounds(Integer competitionId) {
         try {
             if (getAllRounds(competitionId).getRounds().size() != 0)
                 throw new RuntimeException("there are already rounds for this competition");
@@ -58,7 +58,7 @@ public record CompetitionManagementService(ModelMapper mapper, ParticipantReposi
         }
     }
 
-    private CompetitionWithRoundsList tryToGenerateRounds(Long competitionId) {
+    private CompetitionWithRoundsList tryToGenerateRounds(Integer competitionId) {
         Competition competition = competitionRepo.getById(competitionId);
 
         RoundsGenerator generator = new RoundsGenerator(competition);
